+++ example/lib/src/features/client/stories/presentation/pages/stories_modal_screen.dart
@@ -8,10 +8,6 @@ class StoriesModalScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-
-       
-      
-      body: Column(children: []));
+    return Scaffold(body: Column(children: []));
   }
 }